Use a Step Presenter for extra view logic
Instead of grouping methods for different ways to view a Step resource inside the model, break this out into a presenter object. Using this pattern will enable the model to remain concerned with the data management layer and also multiple presenters for HTML, XML etc to be introduced.
