+++ app/models/step.rb
@@ -19,10 +19,6 @@ class Step < ApplicationRecord
       single_date_answer
   end
 
-  def question?
-    contentful_model == "question"
-  end
-
   def primary_call_to_action_text
     return I18n.t("generic.button.next") unless super.present?
     super
