+++ app/presenters/step_presenter.rb
@@ -1,0 +1,5 @@
+class StepPresenter < SimpleDelegator
+  def question?
+    contentful_model == "question"
+  end
+end
