+++ app/views/journeys/show.html.erb
@@ -2,7 +2,7 @@
 
 <h1 class="govuk-heading-xl"><%= @journey.category.capitalize %></h1>
 <dl class="govuk-summary-list">
-  <% @journey.steps.each do |step| %>
+  <% @steps.each do |step| %>
     <% if step.question? %>
       <%= render "#{step.contentful_type}_answer", step: step, answer: step.answer %>
     <% end %>
